feat: option to disable status emoji
Currently, a status emoji is being displayed in the right prompt of
shell. If you have customised your right prompt you may want to disable
the status emoji.

Closes: #64
