+++ functions/_fish_ai_codify.fish
@@ -1,6 +1,10 @@
 #!/usr/bin/env fish
 
 function _fish_ai_codify --description "Turn a comment into a command using AI." --argument-names comment
-    set output (~/.fish-ai/bin/codify "$comment" 2> /dev/null)
+    if test (~/.fish-ai/bin/lookup_setting "debug") = True
+        set output (~/.fish-ai/bin/codify "$comment")
+    else
+        set output (~/.fish-ai/bin/codify "$comment" 2> /dev/null)
+    end
     echo -n "$output"
 end
